characters_db.go の作成
